requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1
click==3.3
django-getenv==1.3.1
itsdangerous==0.24
pycrypto==2.6.1
python-dotenv==0.1.2
pytz==2015.2
requests==2.6.0
six==1.9.0
tbk==0.2.0
